pyproject.toml
[project]
name = "corelinkai-safe"
version = "0.1.0"
description = "Policy/safety agent runtime for A2A policy-compliance benchmarks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
